Address review comments; pass http.Client as an argument, expose the base route and handler via a method, make http/https scheme configurable'
